changes in departments.js

+++ routes/departments.js
@@ -1,13 +1,19 @@
 const express = require("express");
 const router = express.Router();
+const User = require("../models/usersModel");
 const Department = require("../models/departmentsModel");
 const verifyToken = require("../security/verifyToken");
 
 //GET ALL DEPARTMENTS
 router.get("/", verifyToken, async (req, res) => {
   try {
-    const departments = await Department.find();
-    res.status(200).json({message: departments, total: departments.length});
+    const authHeader = req.headers["authorization"];
+    const token = authHeader && authHeader.split(" ")[1];
+    const user = await User.findOne({ token: token })
+    if (!user.access)
+      return res.status(403).json({ message: "Cannot access the content" });
+    const departments = await Department.find({ });
+    res.status(200).json({ message: departments, total: departments.length });
   } catch (err) {
     res.status(500).json({ message: err.message });
   }
@@ -15,14 +21,19 @@ router.get("/", verifyToken, async (req, res) => {
 
 //ADD A DEPARTMENT
 router.post("/", verifyToken, async (req, res) => {
+  const authHeader = req.headers["authorization"];
+  const token = authHeader && authHeader.split(" ")[1];
+  const user = await User.findOne({ token: token })
+  if (!user.admin)
+    return res.status(403).json({ message: "Cannot access the content" });
   const department = new Department({
     name: req.body.name,
     description: req.body.description,
-    organization: req.body.organization,
+    organization: user.organization,
   });
   try {
     const newDepartment = await department.save();
-    res.status(201).json({message: newDepartment});
+    res.status(201).json({ message: newDepartment });
   } catch (err) {
     res.status(400).json({ message: err.message });
   }
@@ -31,11 +42,20 @@ router.post("/", verifyToken, async (req, res) => {
 //GET A SPECIFIC DEPARTMENT
 router.get("/:id", verifyToken, async (req, res) => {
   try {
-    const departments = await Department.findById(req.params.id);
-    if (departments == null) {
+    const authHeader = req.headers["authorization"];
+    const token = authHeader && authHeader.split(" ")[1];
+    const user = await User.findOne({ token: token })
+    if (!user.admin)
+      return res.status(403).json({ message: "Cannot access the content" });
+
+    const department = await Department.findById(req.params.id);
+    if (!department)
       return res.status(404).json({ message: "Cannot find department" });
-    }
-    res.status(200).json({message: departments});
+
+    if (user.organization != department.organization)
+      return res.status(403).json({ message: "Cannot access the content" });
+
+    res.status(200).json({ message: department });
   } catch (err) {
     res.status(500).json({ message: err.message });
   }
@@ -43,37 +63,52 @@ router.get("/:id", verifyToken, async (req, res) => {
 
 //EDIT A SPECIFIC DEPARTMENT
 router.patch("/:id", verifyToken, async (req, res) => {
-  let department;
   try {
-    department = await Department.findById(req.params.id);
-    if (department == null) {
-      return res.status(404).json({ message: "Cannot find department" });
-    }
-  } catch (err) {
-    return res.status(500).json({ message: err.message });
-  }
-  res.department = department;
+    const authHeader = req.headers["authorization"];
+    const token = authHeader && authHeader.split(" ")[1];
+    const user = await User.findOne({ token: token })
+    if (!user.admin)
+      return res.status(403).json({ message: "Cannot access the content" });
 
-  if (req.body.name != null) {
-    department.name = req.body.name;
-  }
-  if (req.body.description != null) {
-    department.description = req.body.description;
-  }
-  if (req.body.organization != null) {
-    department.organization = req.body.organization;
-  }
-  try {
+    const department = await Department.findById(req.params.id);
+    if (!department)
+      return res.status(404).json({ message: "Cannot find department" });
+
+    if (user.organization != department.organization)
+      return res.status(403).json({ message: "Cannot access the content" });
+
+    res.department = department;
+
+    if (req.body.name != null) {
+      department.name = req.body.name;
+    }
+    if (req.body.description != null) {
+      department.description = req.body.description;
+    }
     const updatedDepartment = await res.department.save();
-    res.status(200).json({message: updatedDepartment});
+    res.status(200).json({ message: updatedDepartment });
   } catch (err) {
-    res.status(400).json({ message: err.message });
+    res.status(500).json({ message: err.message });
   }
 });
 
 //DELETE A SPECIFIC DEPARTMENT
 router.delete("/:id", verifyToken, async (req, res) => {
   try {
+    const authHeader = req.headers["authorization"];
+    const token = authHeader && authHeader.split(" ")[1];
+    const user = await User.findOne({ token: token })
+    if (!user.admin)
+      return res.status(403).json({ message: "Cannot access the content" });
+
+    const department = await Department.findById(req.params.id);
+    if (!department)
+      return res.status(404).json({ message: "Cannot find department" });
+
+    if (user.organization != department.organization)
+      return res.status(403).json({ message: "Cannot access the content" });
+
+
     await Department.findByIdAndRemove(req.params.id);
     res.status(200).json({ message: "Department Deleted" });
   } catch (err) {
@@ -82,13 +117,15 @@ router.delete("/:id", verifyToken, async (req, res) => {
 });
 
 //GET DEPARTMENTS FROM ORGANIZATION X
-router.get("/organization/:id", verifyToken, async (req, res) => {
+router.get("/organization", verifyToken, async (req, res) => {
   try {
-    const departments = await Department.find({ organization: req.params.id });
-    if (departments == null) {
-      return res.status(404).json({ message: "Cannot find departments" });
-    }
-    res.status(200).json({message: departments});
+    const authHeader = req.headers["authorization"];
+    const token = authHeader && authHeader.split(" ")[1];
+    const user = await User.findOne({ token: token })
+    if (!user.admin)
+      return res.status(403).json({ message: "Cannot access the content" });
+    const departments = await Department.find({ organization: user.organization });
+    res.status(200).json({ message: departments, total: departments.length });
   } catch (err) {
     res.status(500).json({ message: err.message });
   }
